Move saved games into a separate directory
Saved games should not mix with application code. Make a separate directory
and move it there. And remove absolute path from config files.

To make it work, I had to do some smaller changes about config files and
add a fallback for json loading - because now, some files, like
printColors.txt, can be in the application directory, but saved games
are in different. So if the config is not found as
saved_games/gamename_something.txt, fallbacks to appdir/something.txt.

Signed-off-by: Hades <hades.imagine@gmail.com>